requirements.txt
beautifulsoup4
faiss-cpu
flytekitplugins-deck-standard
gitpython
langchain
langchain-community==0.0.20
langchain-openai
markdownify
mashumaro
openai
playwright
requests
sentence_transformers
tiktoken
torch==2.0.1
unionai==0.1.27
